requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
django-embed-video==1.4.4
django-heroku==0.3.1
django-mongoengine==0.5.4
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
idna==3.3
mongoengine==0.24.1
Pillow==9.2.0
psycopg2==2.9.3
pymongo==3.12.3
pytz==2022.1
requests==2.28.1
sqlparse==0.2.4
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
